+++ packages/indexing/src/history-sync/__tests__/sync-api.test.ts
@@ -331,11 +331,13 @@ describe('Sync API', () => {
         logger
       )
 
+      // @ts-ignore private field
+      sync.defaultStartBlock = 0
       const addSyncJob = jest.fn()
       sync._addSyncJob = addSyncJob as any
 
       const data = { fromBlock: 1, toBlock: 10, models: ['abc123'] }
-      await sync.startModelSync('abc123', data.fromBlock, data.toBlock)
+      await sync.startModelSync('abc123', { startBlock: data.fromBlock, endBlock: data.toBlock })
       expect(addSyncJob).toHaveBeenCalledWith(HISTORY_SYNC_JOB, expect.objectContaining(data))
       expect(Array.from(sync.modelsToSync)).toEqual(data.models)
     })
@@ -350,11 +352,13 @@ describe('Sync API', () => {
         logger
       )
 
+      // @ts-ignore private field
+      sync.defaultStartBlock = 0
       const addSyncJob = jest.fn()
       sync._addSyncJob = addSyncJob as any
 
       const data = { fromBlock: 1, toBlock: 10, models: ['abc123', 'def456'] }
-      await sync.startModelSync(data.models, data.fromBlock, data.toBlock)
+      await sync.startModelSync(data.models, { startBlock: data.fromBlock, endBlock: data.toBlock })
       expect(addSyncJob).toHaveBeenCalledWith(HISTORY_SYNC_JOB, expect.objectContaining(data))
       expect(Array.from(sync.modelsToSync)).toEqual(data.models)
     })