+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.42.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.41.0...@ceramicnetwork/cli@2.42.0-rc.0) (2023-10-11)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
+# [2.41.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.41.0-rc.0...@ceramicnetwork/cli@2.41.0) (2023-10-11)
+
+**Note:** Version bump only for package @ceramicnetwork/cli
+
+
+
+
+
 # [2.41.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/cli@2.40.0...@ceramicnetwork/cli@2.41.0-rc.0) (2023-09-20)
 
 **Note:** Version bump only for package @ceramicnetwork/cli